Add separate backend client id and callback for auth
We can't use the same client id in the frontend and backend - once one
id is approved, the refresh token for the other one is expired.
Therefore, for now use two separate clients.
We can't use the JS app to authenticate the backend client - the auth
library sees the &code=&State= and thinks that these are for it, not for
the backend. Send backend requests directly to the API and have it
redirect back to the frontend after auth exchange